Move service start/stop into WireguardUtilsWindows
