+++ src/daemon/wireguardutils.h
@@ -25,8 +25,8 @@ class WireguardUtils : public QObject {
   virtual ~WireguardUtils() = default;
 
   virtual bool interfaceExists() = 0;
-  virtual bool addInterface() = 0;
-  virtual bool configureInterface(const InterfaceConfig& config) = 0;
+  virtual bool addInterface(const InterfaceConfig& config) = 0;
+  virtual bool updateInterface(const InterfaceConfig& config) = 0;
   virtual bool deleteInterface() = 0;
   virtual peerBytes getThroughputForInterface() = 0;
   virtual bool addRoutePrefix(const IPAddressRange& prefix) = 0;
